+++ src/rich_click/rich_help_formatter.py
@@ -262,6 +262,12 @@ class RichHelpFormatter(click.HelpFormatter):
             from rich.console import COLOR_SYSTEMS
             from rich.segment import Segment
 
+            if self.console.legacy_windows:
+                # Handle legacy windows
+                import colorama  # type: ignore[import-untyped]
+
+                colorama.init()
+
             if self.console.no_color:
                 res = "".join(
                     (